% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_code}
\alias{country_code}
\title{Country code.}
\format{A data frame with three variables:
\describe{
\item{\code{name}}{country name}
\item{\code{code}}{country code to use as \code{fly_from} and \code{fly_to}}
\item{\code{continent}}{continent name}
}}
\usage{
country_code
}
\description{
Contains the \code{name}, ID \code{code}, and \code{continent} of each
country.
}
\keyword{datasets}
