% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{north_america}
\alias{north_america}
\title{Continents' countries.}
\format{A data frame with three variables:
\describe{
  \item{\code{africa}}{codes of countries in Africa}
  \item{\code{asia}}{codes of countries in Asia}
  \item{\code{europe}}{codes of countries in Europe}
  \item{\code{north_america}}{codes of countries in North America}
  \item{\code{oceania}}{codes of countries in Oceania}
  \item{\code{south_america}}{codes of countries in South America}
}}
\usage{
north_america
}
\description{
For each continent, a character vector containing its countries codes.
}
\keyword{datasets}
