% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_flights.R
\name{search_flights}
\alias{search_flights}
\title{Query flight prices.}
\usage{
search_flights(
  fly_from,
  fly_to = "anywhere",
  departure_from = NULL,
  departure_to = NULL,
  sort_by = "POPULARITY",
  price_range = NULL,
  passengers = NULL,
  cabin_class = NULL
)
}
\arguments{
\item{fly_from}{a vector of IDs of the departure locations. E.g.,
`c("City:cordoba_cd_ar", "City:mendoza_md_ar")`.}

\item{fly_to}{a vector of IDs of the arrival destinations.}

\item{departure_from}{search flights from this datetime. If `NULL`, anytime.}

\item{departure_to}{search flights up to this datetime. If `NULL`, anytime.}

\item{sort_by}{one of "POPULARITY", "PRICE", or "QUALITY".}

\item{price_range}{a list with values `min` and/or `max`, setting the price ranges.}

\item{passengers}{If `NULL`, it will use the default: `list(adults = 1, children = 0,`
`infants = 0, adultsHoldBags = 0, adultsHandBags = 0, childrenHoldBags = list(),`
`childrenHandBags = list())`.}

\item{cabin_class}{If `NULL`, it will use the default: `list(cabinClass = "ECONOMY",`
`applyMixedClasses = FALSE)`.}
}
\description{
Query flight prices using the Kiwi API.
}
